Sorted the departure list with timestamp and added the timevalue based on the timezone

+++ app/src/main/java/com/flixbus/timetable/model/Datetime.kt
@@ -8,7 +8,7 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class Datetime(
     @SerializedName("timestamp")
-    val timestamp: Int = 0,
+    val timestamp: Long = 0,
     @SerializedName("tz")
     val tz: String = ""
 ) : Parcelable
